+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+- **Backward compatibility break**: Replaced getException/setException methods with getThrowable/setThrowable in ErrorController.
 - Updated PHP requirement to 7.1.
 
 ## [1.1.0] 2018-04-29